Revert "refactor: split router to edge and lambda (#96)"
This reverts commit b4d56ae4b076a703665954f184af58bc1e9cc76d.
