+++ apps/nextjs/src/app/api/trpc/[trpc]/route.ts
@@ -1,18 +1,20 @@
 import type { NextRequest } from "next/server";
 import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
 
-import { createTRPCContext } from "@acme/api";
-import { lambdaRouter } from "@acme/api/src/lambda";
+import { appRouter, createTRPCContext } from "@acme/api";
 
 // Stripe is incompatible with Edge runtimes due to using Node.js events
 // export const runtime = "edge";
 
 const handler = (req: NextRequest) =>
   fetchRequestHandler({
-    endpoint: "/api/trpc/lambda",
-    router: lambdaRouter,
+    endpoint: "/api/trpc",
+    router: appRouter,
     req: req,
     createContext: createTRPCContext,
+    onError: (opts) => {
+      console.error(opts);
+    },
   });
 
 export { handler as GET, handler as POST };
