--- apps/nextjs/src/app/api/trpc/edge/[trpc]/route.ts
@@ -1,17 +1,0 @@
-import type { NextRequest } from "next/server";
-import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
-
-import { createTRPCContext } from "@acme/api";
-import { edgeRouter } from "@acme/api/src/edge";
-
-export const runtime = "edge";
-
-const handler = (req: NextRequest) =>
-  fetchRequestHandler({
-    endpoint: "/api/trpc/edge",
-    router: edgeRouter,
-    req: req,
-    createContext: createTRPCContext,
-  });
-
-export { handler as GET, handler as POST };
