+++ apps/nextjs/src/trpc/client.ts
@@ -1,10 +1,17 @@
-import { loggerLink } from "@trpc/client";
+"use client";
+
+import { httpBatchLink, loggerLink } from "@trpc/client";
 import { experimental_createTRPCNextAppDirClient } from "@trpc/next/app-dir/client";
 import superjson from "superjson";
 
 import type { AppRouter } from "@acme/api";
 
-import { endingLink } from "./shared";
+const getBaseUrl = () => {
+  if (typeof window !== "undefined") return "";
+  const vc = process.env.VERCEL_URL;
+  if (vc) return `https://${vc}`;
+  return `http://localhost:3000`;
+};
 
 export const api = experimental_createTRPCNextAppDirClient<AppRouter>({
   config() {
@@ -16,7 +23,9 @@ export const api = experimental_createTRPCNextAppDirClient<AppRouter>({
             process.env.NODE_ENV === "development" ||
             (opts.direction === "down" && opts.result instanceof Error),
         }),
-        endingLink(),
+        httpBatchLink({
+          url: `${getBaseUrl()}/api/trpc`,
+        }),
       ],
     };
   },
