+++ apps/nextjs/src/trpc/server.ts
@@ -1,13 +1,18 @@
 "use server";
 
 import { headers } from "next/headers";
-import { loggerLink } from "@trpc/client";
+import { httpBatchLink, loggerLink } from "@trpc/client";
 import { experimental_createTRPCNextAppDirServer } from "@trpc/next/app-dir/server";
 import superjson from "superjson";
 
 import type { AppRouter } from "@acme/api";
 
-import { endingLink } from "./shared";
+const getBaseUrl = () => {
+  if (typeof window !== "undefined") return "";
+  const vc = process.env.VERCEL_URL;
+  if (vc) return `https://${vc}`;
+  return `http://localhost:3000`;
+};
 
 export const api = experimental_createTRPCNextAppDirServer<AppRouter>({
   config() {
@@ -19,8 +24,11 @@ export const api = experimental_createTRPCNextAppDirServer<AppRouter>({
             process.env.NODE_ENV === "development" ||
             (opts.direction === "down" && opts.result instanceof Error),
         }),
-        endingLink({
-          headers: Object.fromEntries(headers().entries()),
+        httpBatchLink({
+          url: `${getBaseUrl()}/api/trpc`,
+          headers: () => {
+            return Object.fromEntries(headers().entries());
+          },
         }),
       ],
     };
