+++ packages/api/src/root.ts
@@ -1,8 +1,13 @@
-import { edgeRouter } from "./edge";
-import { lambdaRouter } from "./lambda";
-import { mergeRouters } from "./trpc";
+import { authRouter } from "./router/auth";
+import { postRouter } from "./router/post";
+import { stripeRouter } from "./router/stripe";
+import { createTRPCRouter } from "./trpc";
 
-// Used to provide a good DX with a single client
-// Then, a custom link is used to generate the correct URL for the request
-const appRouter = mergeRouters(edgeRouter, lambdaRouter);
+export const appRouter = createTRPCRouter({
+  post: postRouter,
+  auth: authRouter,
+  stripe: stripeRouter,
+});
+
+// export type definition of API
 export type AppRouter = typeof appRouter;
