+++ packages/api/src/trpc.ts
@@ -94,7 +94,6 @@ export const t = initTRPC.context<typeof createTRPCContext>().create({
  * @see https://trpc.io/docs/router
  */
 export const createTRPCRouter = t.router;
-export const mergeRouters = t.mergeRouters;
 
 /**
  * Public (unauthed) procedure
