+++ packages/api/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src", "*.ts"]
+  "include": ["src", "index.ts", "transformer.ts"]
 }
